[Área do Professore] Ajustes para mostrar apenas o que corresponde ao professor logado

+++ source/models/MateriasModel.php
@@ -60,6 +60,11 @@ class MateriasModel extends Model
     {
 
         $registro = self::find_by_id(self::getId());
+
+        if(TurmasMateriasModel::find_by_id_materia($registro->id)):
+            return 'erro';
+        endif;
+
         $registro->delete();
 
         return 'ok';
